First set of modifications by Orf. I tried to make the paper more clear to a native-English-speaking audience. The English is still pretty shaky in many parts, however, and could use another serious edit. I was not not comforatable changing some parts because I was afraid of changing the intended meaning, which sometimes was not clear to me.
I am concerned about the lack of the passive voice throughout this
paper. It is not clear to me that the word "We" should appear anywhere!
Please look into whether the active voice is allowed for this
publication. Changing completely over to the passive voice will take a
considerable effort but we want the paper to be accepted.

There are incomplete references (missing the journal/conference) as well
as capitalization issues within the references which I have not changed.